+++ src/main/java/proxyauth/ProxyListener.java
@@ -1,6 +1,29 @@
+/*
+ * This file is part of ProxyAuth - https://github.com/Zeckie/ProxyAuth
+ * ProxyAuth is Copyright (c) 2021 Zeckie
+ *
+ * ProxyAuth is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free
+ * Software Foundation, version 3.
+ *
+ * ProxyAuth is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ *  for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with ProxyAuth. If you have the source code, this is in a file called
+ * LICENSE. If you have the built jar file, the licence can be viewed by
+ * running "java -jar ProxyAuth-<version>.jar licence".
+ * Otherwise, see <https://www.gnu.org/licenses/>.
+ */
+
 package proxyauth;
 
+import proxyauth.conf.Configuration;
+
 import java.io.IOException;
+import java.net.InetAddress;
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.util.LinkedHashSet;
@@ -10,7 +33,6 @@ import java.util.Set;
  * Listen for requests and create threads to handle them
  *
  * @author Zeckie
- * Copyright and licence details in Main.java
  */
 public class ProxyListener implements Runnable, StatusListener<ProxyRequest> {
 
@@ -25,12 +47,26 @@ public class ProxyListener implements Runnable, StatusListener<ProxyRequest> {
     private final Set<ProxyRequest> activeRequests = new LinkedHashSet<>();
 
 
+    public final Configuration config;
+
+    public ProxyListener(Configuration configuration) {
+        config = configuration;
+    }
+
+
     @Override
     public void run() {
 
-        try (ServerSocket incoming = new ServerSocket(Configuration.LISTEN_PORT, Configuration.LISTEN_BACKLOG, Configuration.listenAddress)) {
+        try (
+                ServerSocket incoming = new ServerSocket(
+                        config.LISTEN_PORT.getValue(),
+                        config.LISTEN_BACKLOG.getValue(),
+                        InetAddress.getByName(config.LISTEN_ADDRESS.getValue())
+                )
+        ) {
             System.out.println("Listening " + incoming);
 
+            //noinspection InfiniteLoopStatement (CTRL+C to stop)
             while (true) {
                 Socket sock = incoming.accept();
 
@@ -39,7 +75,7 @@ public class ProxyListener implements Runnable, StatusListener<ProxyRequest> {
                     activeRequests.add(proxyRequest);
                     activeRequests.notifyAll();
                     proxyRequest.start();
-                    while (activeRequests.size() >= Configuration.MAX_ACTIVE_REQUESTS) {
+                    while (activeRequests.size() >= config.MAX_ACTIVE_REQUESTS.getValue()) {
                         System.out.println("Active request limit reached - waiting for a request to finish");
                         activeRequests.wait();
                     }
@@ -60,6 +96,10 @@ public class ProxyListener implements Runnable, StatusListener<ProxyRequest> {
                     "Active requests:" + activeRequests.size() + "\n" +
                     "Active threads:" + THREADS.activeCount()
             );
+
+            if (config.DEBUG.getValue()) {
+                THREADS.list();
+            }
 
             activeRequests.notifyAll();
         }